// client/src/redux/posts/posts.types.ts
export enum PostsTypes {
  FETCH_POSTS_BY_USERNAME_START = "FETCH_POSTS_BY_USERNAME_START",
  FETCH_POSTS_START = "FETCH_POSTS_START",
  FETCH_POSTS_SUCCESS = "FETCH_POSTS_SUCCESS",
  FETCH_POSTS_FAILURE = "FETCH_POSTS_FAILURE",
  DELETE_POST_START = "DELETE_POST_START",
  DELETE_POST_SUCCESS = "DELETE_POST_SUCCESS",
  DELETE_POST_FAILURE = "DELETE_POST_FAILURE",
  ADD_POST_START = "ADD_POST_START",
  ADD_POST_SUCCESS = "ADD_POST_SUCCESS",
  ADD_POST_FAILURE = "ADD_POST_FAILURE",
  SET_DELETE_POST_CONFIRMATION_DIALOG_VISIBILITY = "SET_DELETE_POST_CONFIRMATION_DIALOG_VISIBILITY",
  LIKE_POST_START = "LIKE_POST_START",
  LIKE_POST_SUCCESS = "LIKE_POST_SUCCESS",
  LIKE_POST_FAILURE = "LIKE_POST_FAILURE",
  FETCH_SELECTED_POST_START = "FETCH_SELECTED_POST_START",
  FETCH_SELECTED_POST_SUCCESS = "FETCH_SELECTED_POST_SUCCESS",
  FETCH_SELECTED_POST_FAILURE = "FETCH_SELECTED_POST_FAILURE",
  RESET_POSTS_DATA = "RESET_POSTS_DATA",
}